final commit

+++ app/Http/Controllers/AlbumController.php
@@ -6,6 +6,7 @@ use App\Models\Album;
 use App\Http\Resources\V1\AlbumResource;
 use App\Http\Requests\StoreAlbumRequest;
 use App\Http\Requests\UpdateAlbumRequest;
+use Illuminate\Http\Request;
 use PhpParser\Node\Stmt\TryCatch;
 
 class AlbumController extends Controller
@@ -15,9 +16,9 @@ class AlbumController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        return AlbumResource::collection(Album::paginate());
+        return AlbumResource::collection(Album::where('user_id',$request->user()->id)->paginate());
     }
 
     /**
@@ -29,7 +30,9 @@ class AlbumController extends Controller
     public function store(StoreAlbumRequest $request)
     {
         try {
-        $album = Album::create($request->all());
+            $data = $request->all();
+            $data['user_id'] = $request->user()->id;
+            $album = Album::create($data);
 
         return new AlbumResource($album) ; 
     
@@ -50,8 +53,11 @@ class AlbumController extends Controller
      * @param  \App\Models\Album  $album
      * @return \Illuminate\Http\Response
      */
-    public function show(Album $album)
+    public function show(Request $request,Album $album)
     {
+        if ($request->user()->id != $album->user_id) {
+            return abort(403,'unauthorize request');
+        }
         return new AlbumResource($album) ;
     }
 
@@ -65,6 +71,9 @@ class AlbumController extends Controller
     public function update(UpdateAlbumRequest $request, Album $album)
     {
         try {
+            if ($request->user()->id != $album->user_id) {
+                return abort(403,'unauthorize request');
+            }
             $album -> update($request->all());
 
             return response()->json([
@@ -88,8 +97,11 @@ class AlbumController extends Controller
      * @param  \App\Models\Album  $album
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Album $album)
+    public function destroy(Request $request,Album $album)
     {
+        if ($request->user()->id != $album->user_id) {
+            return abort(403,'unauthorize request');
+        }
         $album->delete();
 
         return response()->json([
